requirements.txt
requests
pyyaml
ujson
bs4
lxml
